refactor(stocks): background generation saves report directly, no callback HTTP hop
Previously the in-process background path POSTed the LLM result back to the
save-report-callback HTTP route. On the same server that round-trip is
unnecessary, so call the save logic directly.

- Extract the entire save-report-callback body into a shared util
  saveTickerReportAndAdvanceGeneration() (save-report-callback-utils.ts):
  revalidation decision -> switch/store per report type -> mark step completed
  + clear inProgressStep/lastInvocationTime -> triggerGenerationOfAReportSimplified.
  The HTTP route now just parses the request and calls it, so the lambda path
  and the background path run byte-for-byte identical save + chaining logic.
- Move the background processor out of llm-callback-lambda-utils into its own
  file (background-llm-generation-utils.ts) as processStockReportLLMResponseInBackground:
  run getLLMResponse in-process, then call saveTickerReportAndAdvanceGeneration
  directly instead of fetch()-ing the callback URL.
- llm-callback-lambda-utils now passes explicit context (symbol/exchange/
  reportType/generationRequestId/invocation/prompt/schema) to the background
  processor; still gated on USE_LAMBDA_FOR_LLM_RESPONSE and scoped to stocks
  (reportType + exchange present). Daily movers keep using the lambda.
